Adding the sub handlers for prelogin :
